Added EchoServer.java and EchoClient.java
I got these two files from a YouTube tutorial on a multi-threaded echo server. We would need to change some variable and class names to fit the requirements, as well as implement a way so that the server only takes one character at a time from the client, but I figured this gives us something to start off with.